Try: Reduced specificity base block margins.
This is a simple change that benefits themes. It reduces the specificity of the rule that provides a baseline block margin.

The baseline block margin is the margin above and below every block, which these blocks are born with.

In master, this rule is so specific that it overrides many editor styles that should work.